Tests project, inside code documentation is updated, README.md updated, Doxyfile added, Documentation auto generation enabled.
