+++ RFPPortalWebsite/RFPPortalWebsite.csproj
@@ -4,6 +4,7 @@
     <UserSecretsId>06fcddd3-c0ec-4ab4-ab56-49cbfd37961d</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerComposeProjectPath>..\docker-compose.dcproj</DockerComposeProjectPath>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.10" />
